made separate file with writeToDocumentDirectory FileManager extension
